Rename selector module to axeswidgets
